switch to enable strict search, default: False
Add a "strict" switch to our search functions,
still defaulting to fuzzy searches.

Currently search fields are concatenated with spaces,
which default to the OR operator.
The values for the field are grouped with parentheses.
This Lucene syntax results in a fuzzy search,
where you also get results that just match part of the query,
but the close results are still the first results.

This is nice if you want to find additional information
and you are happy that you get results even if the query is not 100 %
accurate.

However, sometimes you don't want any results that don't
fit exactly what you asked for.
You want to know how many items match your query exactly
without filtering the results for scores and mismatches.

When "strict" is set to True then every field value
is enclosed in quotation marks, forming a lucene phrase.
That whole phrase has to match.
Additionally the fields are concatenated with AND.

Alternatives:
It would also be possible to prepend a + (plus) to every word
of the field values.
This would force every word to match, but any order is allowed.

Please note, that you can use the query parameter directly
for full Lucene support.

Signed-off-by: Johannes Dewender <github@JonnyJD.net>