use search fields as phrases, not groups
Currently each search field value is enclosed in parentheses,
effectively making it a group.
That is enough to make sure every word is searched in the context of
this field.
Additionally, quotation marks are stripped from the values of fields.

However, when you want to search for artist="The White Stripes"
you expect that get results including the whole artist name you
provided, not just one of these words.
Note, that (The White Stripes) is the same as (The OR White OR Stripes).

You could also prepend a plus sign to every word:
(+The +White +Stripes).
That would effectively search any permutation of these words.

In this commit the simple "enclose in parentheses" is implemented,
having a strict search for an order.

Please note that you can still use the query parameter to do fancy
stuff.

Signed-off-by: Johannes Dewender <github@JonnyJD.net>