prevent http 400 in search queries
If fields are not set, we don't change the query.
Full Lucene syntax is available in the query
and http status 400 messages can occur.
The user is fully responsible to prevent this.

If there are any fields set,
we escape any lucene special characters that might appear in the query.

In fuzzy mode (strict=False, default) we lowercase the field values,
and possibly the query (if fields are set).

In strict mode we enclose the field values,
and possibly the query (if fields are set) in quotation marks.

Both changes should prevent uppercase AND / OR from appearing in the query.
Otherwise, these operators could lead to http 400 messages
when placed incorrectly.

Alternatives:

Always escape lucene special chars in the query in strict mode,
even if no fields are set.
That would support doing a plain search including entities with AND
or special characters.
However, even the MB website just fails when you search for AND.
Advanced queries would only be possible in fuzzy mode,
but this is the default mode anyways.

Signed-off-by: Johannes Dewender <github@JonnyJD.net>