Make the codebase support python2 (>=.6) and 3
This is done by moving all imports that are not the same in both python2
and 3 (2to3 was used to get the list of the affected imports) into a
compat module.

All tests pass in python 2.6, 2.7, 3.2.2 :sparkles:

The examples do NOT work in python 3 because io.StringIO can't deal with
bytes, but Johannes Dewender has a solution in review which can solve
that issue.

Signed-off-by: Wieland Hoffmann <themineo@gmail.com>