Fix error in make_artist_credit when artist names are overridden
Certain Musicbrainz releases credit artists differently than their true names.
A good example of this can be found here, where a album by Shira Kammen and Pam
Swan is credited to "Kammen & Swan".

http://musicbrainz.org/release/9e323e0c-adb0-4c4f-8eb7-75f19503d7e9

The XML output records these overrides by slipping a "name" object outside the
artist object itself. It looks like this in json:

        "artist-credit": [
            {
                "name": "Kammen",
                "artist": {
                    "sort-name": "Kammen, Shira",
                    "id": "3736f006-1607-42e1-ba50-47c5090d68d0",
                    "name": "Shira Kammen"
                }
            },
            " & ",
            {
                "name": "Swan",
                "artist": {
                    "sort-name": "Swan, Pam",
                    "id": "7c92f4c4-9522-4c47-aad4-3bcf718e75d7",
                    "name": "Pam Swan"
                }
            }
        ],

But make_artist_credit produced this output, which is wrong:

        "artist-credit-phrase": "Shira Kammen & Pam Swan"

I added a test in the function for the extra name object, and use it instead
of artist:name if it exists. The function now generates something that matches
the website output.

        "artist-credit-phrase": "Kammen & Swan"

Signed-off-by: Galen Hazelwood <galenhz@gmail.com>