Add release-list element to parse_work and parse_recording
When retrieving a release with the "recording-level-rels" and "work-level-rels"
includes, the server will return relation-lists for recordings and works. But
because the code does not expect these elements to be present, they do not get
parsed and the information is lost. Added these potential elements to the list
so they will be parsed if present.

Signed-off-by: Galen Hazelwood <galenhz@gmail.com>