added travis-ci support, update tox config
travis-ci:
Needs an install, otherwise travis tries rake/requirements.txt
Python 3.3 is the latest, we should test for this.
Previous Python 3 versions are not as important.
The base version, Python 2.6 is important though.
We should check every branch and also pull-requests.
So we just exclude branches were we know we will fail (gh-pages, old).

tox:
Remove pypy (having it in travis should be enough)
Remove 3.1 (old), add 3.3 latest
Not even sure if tox.ini should be in the repo or rather
having a local one (with the versions installed) is better.

Adding travis-ci support was proposed by Sam Doshi.

Signed-off-by: Johannes Dewender <github@JonnyJD.net>