fix _check_filter_and_make_params() usage
The function wants an entitiy as first argument.
The usage is correct in single entitities, but wasn't in entity lists.

The problem was introduced in
86110dc2886a74d85dc349d11a2f7672e34ba035
where the function parameters were changed.
That gave
TypeError: ... takes at least 2 arguments (2 give)
when using get_releases_by_discid() and others.

The problem was hidden with the fix in
4260ed51273babb909a27fe07d68923158d2e9ec
No TypeError anymore,
but the entity was set to the includes, includes were empty in return.

For single entity requests this was fixed correctly in
088562d764db85355fb075c30d6b5590a804bb2e

This now finally fixes this for entity list requests.