doc: read supported search fields from code
There is only one way to make sure code and documentation are in sync:
The documentated fields should be read from the code.
Since I couldn't find a way to evaluate a python construct in the
docstring and I didn't want to introduce a constant for every entity,
I just marked and included the relevant parts of the code.