document browsing usage
This example explains paging and tries to persuade the user to filter
the query whenever possible.

The example will only fetch 2 pages (168 releases).
That is bigger than the highest limit, so paging is necessary,
but won't strain the server too much when the examples are tried.

This needs the fix to filter by releases by type.
(merged in this branch)

Signed-off-by: Johannes Dewender <github@JonnyJD.net>