submission: move list conversion to api functions
Converting single items to lists is user input transformation and this
should be done directly in the API functions (or at least be visible
from there).

While removing this conversion in the backend,
I also changed the names of a couple of variables so one can guess the
number of entities envolved.
recording2isrcs gives multiple isrcs for each recording.
release2barcode gives only one barcode for every release.
Just "puids" doesn't tell much about what the variable holds.

The docstrings were also updated so they all use the same syntax for the
hint on what kind of dict they expect.
I documented that isrcs and co. can be give as lists or single values.

I also renamed variables with only 1 letter.
"p" is not a good variable name for a puid.

We don't have to test if a rating is "int".
str(rating) will make it an str, no matter if it was int, str
or unicode before. (these shouldn't have encoding problems)

Again, I change the indentation to 4 spaces in the functions involved.

Signed-off-by: Johannes Dewender <github@JonnyJD.net>