parse release event lists
With the new area changes, there is an additional list of release events,
which each have a date and an area.
All previous release information including one date
(which is the arliest from the event list) and a country string
are still provided at the usual place directly attached to release
entities.

Signed-off-by: Johannes Dewender <github@JonnyJD.net>