DeprecationWarning -> Warning, remove simplefilter
Enabling DeprecationWarnings can only be done application-wide and not
only for this library. So our line had an unwanted impact on the
importing application. (see #157)

We remove that line and use plain Warnings for PUID and echoprint
deprecation.
The reasoning is that this is not something that is deprecated and
otherwise (still) works fine, but something that is already broken, even
though still part of the API.

Signed-off-by: Johannes Dewender <github@JonnyJD.net>