Use singular to refer to valid browse includes
This means that we can get the valid includes in _browse_impl
based on the entity type instead of having to pass it in.

Add tests to check each combination of browse/parameters, including
type/status, offset/limit.

Also added a test to check that browse includes are a subset of
GET includes, but this is failing currently due to it not being
the case. This is fixed by 381349ad9.

Signed-off-by: Alastair Porter <alastair@porter.net.nz>