Don't try and import fixtag from ElementTree
We had a comment suggesting that fixtag didn't exist in python 2.6,
but in fact it's the other way around. It exists in 2.6, but only
in the python implementation (not the C implementation), and doesn't
exist in any more recent version. Because of this, just define the
function and use it in all versions.

Signed-off-by: Alastair Porter <alastair@porter.net.nz>