Add missing valid seach parameters
We were missing search parameters for event, instrument, and place which
means that any query for these items with a named field would fail.
Also missing a few (new?) fields for some existing types.
Added tests for all searches to ensure that named fields work
as expected, and errors for invalid fields work.

Signed-off-by: Alastair Porter <alastair@porter.net.nz>