Add genres to TAG_INCLUDES
Genres were introduced in MusicBrainz a few months ago:
https://blog.metabrainz.org/2018/11/02/musicbrainz-introducing-genres/

One thing to note in that post:
> For now genres are exactly the same as (folksonomy) tags behind the scenes

An implication of this is that the new `genres` and `user-genres` are
applicable for all the cases where `tags` and `user-tags` are. Thus it
seemed like the "path of least resistance" to simply add these to the
potential include parameters for tags.

See also my question (and its answer(s)) on IRC:
https://chatlogs.metabrainz.org/brainzbot/metabrainz/msg/4336139/

Signed-off-by: Frederik “Freso” S. Olesen <freso.dk@gmail.com>