Specify supported version(s) of Python in setup.py
This specifies which versions of Python the library is compatible with.
This both prevents the module from being installed in an incompatible
Python environment, and will also enable users after future releases to
still `pip install musicbrainzngs` and get a version that is compatible
with their Python setup.

Versions are based on the ones currently being tested for in Travis
according to `.travis.yml`.

See https://packaging.python.org/guides/distributing-packages-using-setuptools/#python-requires
and https://packaging.python.org/guides/dropping-older-python-versions/

Signed-off-by: Frederik “Freso” S. Olesen <freso.dk@gmail.com>